refactor: move functions centered in run.py for public api to registry.py and results.py.
Prioritizes separation of concerns with public api determined in __init__.py
