+++ src/workrb/__init__.py
@@ -3,13 +3,9 @@ WorkRB - A benchmarking framework for evaluating models on various tasks.
 """
 
 from workrb import data, metrics, models, tasks
-from workrb.run import (
-    evaluate,
-    evaluate_multiple_models,
-    get_tasks_overview,
-    list_available_tasks,
-    load_results,
-)
+from workrb.registry import list_available_tasks
+from workrb.results import load_results
+from workrb.run import evaluate, evaluate_multiple_models, get_tasks_overview
 
 __all__ = [
     "data",
